Improved stop word functionality
* Expanded capability of the stop word to be executed at any point during an executed assistant pipeline.
* Stop word now has the ability to halt media playback (Streaming URLs, Playlists, Single files) acting like a Stop button. This will only occur when there are no active assist pipelines running. I.E. if a wake word is triggered during media playback, the audio will be ducked and the assistant pipeline will run normally without stopping media playback. Once complete, audio will be un-ducked.
* Added a soft internal stop word probability filter to prevent false positives.
* Added detailed "debug" logging.

NOTE: We need a better stop word model. After some tweaking similar words like steep and step can sometimes trigger, and words like start, studder and stair rarely triggers the stop word. I've raised the internal prob thresh to .8 and the sliding window size to 7 and that seems to have cut down on the amount of false positives but a value of .85 seems to prevent the actual word stop from working some of the time as well as saying the stop word in an annoying way like "stoooop" but I'd rather not have to repeat the stop word more than once so we don't want to make it too aggressive.

Testing environment: I spent another several hours tweaking my audio stack to make sure noise cancelation was working as best it could be in my enclosure. I still need to move the microphones so for me, I have to sometimes reduce the speaker volume a bit before the assistant picks up wake words or stop words. Just a note of caution, if your audio stack isn't working correctly, this WILL affect how stop and wake words are triggered even if it seems TTS is working perfectly. This is one of the major reasons I spent some extra time to make sure the "Wake word sensitivity" drop down setting works as intended and that the prob filters were correctly implemented for the micro and openWakeWord libs. If you're having issues getting the wake word to trigger, play with this setting.